Move user list adapter to adapters folder
