+++ app/src/main/AndroidManifest.xml
@@ -40,7 +40,8 @@
 
         </activity>
 
-        <activity android:name=".Activities.ListUserActivity"/>
+        <activity android:name=".Activities.ListUserActivity"
+            android:exported="true"/>
 
         <activity android:name=".Activities.ProfileActivity"
             android:exported="true">
